test(formatters): simplify helper graph in integration suite
