+++ tests/integration/formatters/format_contract_tests.py
@@ -11,6 +11,12 @@ Format contracts ensure:
 4. Cross-format data integrity
 """
 
+import csv
+import io
+from dataclasses import dataclass
+import re
+from pathlib import Path
+import tempfile
 from typing import Any
 
 import pytest
@@ -19,33 +25,886 @@ from tree_sitter_analyzer.mcp.tools.analyze_code_structure_tool import (
     AnalyzeCodeStructureTool as TableFormatTool,
 )
 
-from ._format_contract_tests_helpers import (
-    assert_access_modifier_consistency,
-    assert_backward_compatibility_contract,
-    assert_cross_format_data_integrity,
-    assert_csv_parameter_encoding,
-    assert_field_count_consistency,
-    assert_line_number_consistency,
-    assert_method_count_consistency,
-    assert_outputs_contain_class_name,
-    cleanup_comprehensive_contract_fixture,
-    collect_format_outputs,
-    create_comprehensive_contract_fixture,
-    extract_compact_format_info,
-    extract_contract_infos,
-    extract_csv_format_info,
-    extract_full_format_info,
-    parse_markdown_table,
-    validate_class_consistency,
-    validate_compact_format_contracts,
-    validate_count_consistency,
-    validate_cross_format_contracts,
-    validate_csv_format_contracts,
-    validate_csv_parameters,
-    validate_field_consistency,
-    validate_full_format_contracts,
-    validate_method_consistency,
-)
+FORMAT_TYPES = ("full", "compact", "csv")
+
+CSV_HEADER = [
+    "Type",
+    "Name",
+    "ReturnType",
+    "Parameters",
+    "Access",
+    "Static",
+    "Final",
+    "Line",
+]
+
+
+@dataclass(frozen=True)
+class _CountExpectation:
+    key: str
+    label: str
+    expected_count: int
+
+
+def extract_table_output(result: dict[str, Any]) -> str:
+    """Return table output from a tool response regardless of response shape."""
+    return result["table_output"] if "table_output" in result else result.get("content", "")
+
+
+async def collect_format_outputs(
+    tool: Any,
+    file_path: Path,
+    format_types: tuple[str, ...] = FORMAT_TYPES,
+) -> dict[str, str]:
+    """Generate table outputs for the requested format types."""
+    outputs = {}
+    for format_type in format_types:
+        result = await tool.execute(
+            {
+                "file_path": str(file_path),
+                "format_type": format_type,
+                "language": "java",
+                "output_format": "json",
+            }
+        )
+        outputs[format_type] = extract_table_output(result)
+    return outputs
+
+
+def extract_contract_infos(
+    contract_validator: Any, outputs: dict[str, str]
+) -> tuple[dict[str, Any], dict[str, Any], dict[str, Any]]:
+    """Extract full, compact, and CSV contract information."""
+    return (
+        contract_validator._extract_full_format_info(outputs["full"]),
+        contract_validator._extract_compact_format_info(outputs["compact"]),
+        contract_validator._extract_csv_format_info(outputs["csv"]),
+    )
+
+
+def assert_outputs_contain_class_name(outputs: dict[str, str], class_name: str) -> None:
+    """Assert every generated format includes the source class name."""
+    for format_type, output in outputs.items():
+        assert class_name in output, f"{format_type} format missing class name"
+
+
+def assert_method_count_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    expected_methods: int,
+) -> None:
+    """Assert all formats report the same method count."""
+    _assert_named_count_consistency(
+        full_info,
+        compact_info,
+        csv_info,
+        _CountExpectation("methods", "Method", expected_methods),
+    )
+
+
+def assert_field_count_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    expected_fields: int,
+) -> None:
+    """Assert all formats report the same field count."""
+    _assert_named_count_consistency(
+        full_info,
+        compact_info,
+        csv_info,
+        _CountExpectation("fields", "Field", expected_fields),
+    )
+
+
+def _assert_named_count_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    expectation: _CountExpectation,
+) -> None:
+    full_count = len(full_info.get(expectation.key, []))
+    compact_count = len(compact_info.get(expectation.key, []))
+    csv_count = len(csv_info.get(expectation.key, []))
+
+    assert full_count == compact_count, (
+        f"{expectation.label} count mismatch: "
+        f"full={full_count}, compact={compact_count}"
+    )
+    assert full_count == csv_count, (
+        f"{expectation.label} count mismatch: full={full_count}, csv={csv_count}"
+    )
+    assert full_count >= expectation.expected_count - 1, (
+        f"Expected at least {expectation.expected_count - 1} "
+        f"{expectation.key}, got {full_count}"
+    )
+
+
+def assert_csv_parameter_encoding(csv_output: str, contract_validator: Any) -> None:
+    """Assert CSV method parameter strings follow the expected encoding."""
+    try:
+        reader = csv.reader(io.StringIO(csv_output))
+        next(reader)
+        for row in reader:
+            if len(row) >= 4 and row[0] in ["method", "constructor"]:
+                _assert_valid_parameter_cell(row[3], contract_validator)
+    except (csv.Error, StopIteration) as exc:
+        raise AssertionError("Failed to parse CSV output for parameter validation") from exc
+
+
+def _assert_valid_parameter_cell(parameters: str, contract_validator: Any) -> None:
+    if parameters.strip():
+        assert contract_validator._validate_csv_parameters(parameters), (
+            f"Invalid parameter encoding: {parameters}"
+        )
+
+
+def assert_access_modifier_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+) -> None:
+    """Assert method access modifiers match across formats."""
+    for method_row in full_info.get("methods", []):
+        if len(method_row) < 4:
+            continue
+
+        method_name = method_row[0]
+        full_access = method_row[3]
+        compact_method = _find_markdown_row(
+            compact_info.get("methods", []), method_name, minimum_length=3
+        )
+        if compact_method:
+            _assert_matching_access(
+                method_name, full_access, compact_method[2], "compact"
+            )
+
+        csv_method = _find_csv_row(csv_info.get("methods", []), method_name)
+        if csv_method:
+            _assert_matching_access(
+                method_name, full_access, csv_method.get("Access", ""), "csv"
+            )
+
+
+def assert_line_number_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+) -> None:
+    """Assert method line numbers match across formats."""
+    for method_row in full_info.get("methods", []):
+        if len(method_row) < 5:
+            continue
+
+        method_name = method_row[0]
+        full_line = method_row[4]
+        compact_method = _find_markdown_row(
+            compact_info.get("methods", []),
+            method_name,
+            minimum_length=4,
+            line_number=full_line,
+        )
+        if compact_method:
+            _assert_matching_line(method_name, full_line, compact_method[3], "compact")
+
+        csv_method = _find_csv_row(
+            csv_info.get("methods", []), method_name, line_number=full_line
+        )
+        if csv_method:
+            _assert_matching_line(
+                method_name, full_line, csv_method.get("Line", ""), "csv"
+            )
+
+
+def _find_markdown_row(
+    rows: list[list[str]],
+    name: str,
+    minimum_length: int,
+    line_number: str | None = None,
+) -> list[str] | None:
+    for row in rows:
+        if len(row) < minimum_length or row[0] != name:
+            continue
+        if line_number is None or row[minimum_length - 1] == line_number:
+            return row
+    return None
+
+
+def _find_csv_row(
+    rows: list[dict[str, str]], name: str, line_number: str | None = None
+) -> dict[str, str] | None:
+    for row in rows:
+        if row.get("Name") != name:
+            continue
+        if line_number is None or row.get("Line", "") == line_number:
+            return row
+    return None
+
+
+def _assert_matching_access(
+    method_name: str, full_access: str, compared_access: str, compared_format: str
+) -> None:
+    assert full_access == compared_access, (
+        f"Access modifier mismatch for method {method_name}: "
+        f"full='{full_access}', {compared_format}='{compared_access}'"
+    )
+
+
+def _assert_matching_line(
+    method_name: str, full_line: str, compared_line: str, compared_format: str
+) -> None:
+    assert full_line == compared_line, (
+        f"Line number mismatch for method {method_name}: "
+        f"full='{full_line}', {compared_format}='{compared_line}'"
+    )
+
+
+def assert_backward_compatibility_contract(current_outputs: dict[str, str]) -> None:
+    """Assert the legacy format sections are still present."""
+    for section in ["# ", "## Class Info", "## Methods", "## Fields"]:
+        assert section in current_outputs["full"], (
+            f"Backward compatibility: missing {section} in full format"
+        )
+
+    for section in ["# ", "## Info", "## Methods", "## Fields"]:
+        assert section in current_outputs["compact"], (
+            f"Backward compatibility: missing {section} in compact format"
+        )
+
+    expected_csv_header = "Type,Name,ReturnType,Parameters,Access,Static,Final,Line"
+    assert current_outputs["csv"].startswith(expected_csv_header), (
+        "Backward compatibility: CSV header format changed"
+    )
+
+
+def assert_cross_format_data_integrity(
+    outputs: dict[str, str],
+    class_name: str,
+    is_consistent: bool,
+    is_contract_valid: bool,
+    report: dict[str, Any],
+) -> None:
+    """Assert all cross-format data integrity checks pass."""
+    assert is_consistent and is_contract_valid, (
+        f"Data integrity contract violated: {report['violations']}"
+    )
+    assert_outputs_contain_class_name(outputs, class_name)
+
+    full_lines = len(outputs["full"].split("\n"))
+    compact_lines = len(outputs["compact"].split("\n"))
+    assert full_lines >= compact_lines, (
+        "Full format should have more or equal lines compared to compact format"
+    )
+
+
+def validate_class_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    violations: list[str],
+) -> bool:
+    """Validate class information consistency."""
+    full_class = full_info.get("class_name", "")
+    compact_class = compact_info.get("class_name", "")
+    csv_class = csv_info.get("class_name", "")
+
+    _append_class_mismatch(violations, "full", full_class, "compact", compact_class)
+    _append_class_mismatch(violations, "full", full_class, "csv", csv_class)
+    _append_class_mismatch(violations, "compact", compact_class, "csv", csv_class)
+    return True
+
+
+def validate_count_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    violations: list[str],
+) -> bool:
+    """Validate count information consistency."""
+    del csv_info
+    full_method_count = len(full_info.get("methods", []))
+    full_field_count = len(full_info.get("fields", []))
+    compact_counts = compact_info.get("counts", {})
+
+    _append_reported_count_mismatch(
+        violations,
+        compact_counts.get("methods"),
+        full_method_count,
+        "methods",
+    )
+    _append_reported_count_mismatch(
+        violations,
+        compact_counts.get("fields"),
+        full_field_count,
+        "fields",
+    )
+    return True
+
+
+def validate_full_format_contracts(output: str, violations: list[str]) -> bool:
+    """Validate Full Format specific contracts."""
+    for section in ["## Class Info", "## Methods", "## Fields"]:
+        if section not in output:
+            violations.append(f"Full format missing required section: {section}")
+
+    _validate_section_contains(
+        output,
+        "Methods",
+        ["Parameters"],
+        "Full format Methods table missing {column} column",
+        violations,
+    )
+    _validate_section_contains(
+        output,
+        "Fields",
+        ["Static", "Final"],
+        "Full format Fields table missing {column} column",
+        violations,
+    )
+    return True
+
+
+def validate_compact_format_contracts(
+    output: str, violations: list[str], warnings: list[str]
+) -> bool:
+    """Validate Compact Format specific contracts."""
+    if "## Info" not in output:
+        violations.append("Compact format missing required Info section")
+
+    if "Parameters" in output:
+        warnings.append("Compact format should omit detailed parameter information")
+
+    header_match = re.search(r"^# (.+)$", output, re.MULTILINE)
+    if header_match and "." in header_match.group(1):
+        warnings.append("Compact format header should omit package information")
+
+    return True
+
+
+def validate_csv_format_contracts(output: str, violations: list[str]) -> bool:
+    """Validate CSV Format specific contracts."""
+    try:
+        reader = csv.reader(io.StringIO(output))
+        header = next(reader)
+        rows = list(reader)
+    except (csv.Error, StopIteration) as exc:
+        violations.append(f"CSV format structure violation: {exc}")
+        return True
+
+    if header != CSV_HEADER:
+        violations.append(f"CSV format header contract violation: {header}")
+
+    if not any(row and row[0] == "class" for row in rows):
+        violations.append("CSV format must contain at least one class row")
+
+    _validate_csv_parameter_rows(rows, violations)
+    return True
+
+
+def validate_cross_format_contracts(
+    full_output: str, compact_output: str, violations: list[str]
+) -> bool:
+    """Validate contracts that span multiple formats."""
+    full_info = extract_full_format_info(full_output)
+    compact_info = extract_compact_format_info(compact_output)
+
+    _append_extra_compact_entries(
+        violations, full_info, compact_info, "methods", "methods"
+    )
+    _append_extra_compact_entries(
+        violations, full_info, compact_info, "fields", "fields"
+    )
+    return True
+
+
+def validate_csv_parameters(parameters: str) -> bool:
+    """Validate CSV parameter encoding format."""
+    if not parameters.strip():
+        return True
+
+    return all(_is_valid_csv_parameter(param) for param in parameters.split(";"))
+
+
+def _append_class_mismatch(
+    violations: list[str],
+    left_label: str,
+    left_class: str,
+    right_label: str,
+    right_class: str,
+) -> None:
+    if left_class and right_class and left_class != right_class:
+        violations.append(
+            f"Class name mismatch: "
+            f"{left_label}='{left_class}', {right_label}='{right_class}'"
+        )
+
+
+def _append_reported_count_mismatch(
+    violations: list[str],
+    reported_count: int | None,
+    actual_count: int,
+    label: str,
+) -> None:
+    if reported_count is not None and reported_count != actual_count:
+        violations.append(
+            f"Compact format reports {reported_count} {label}, "
+            f"but full format has {actual_count}"
+        )
+
+
+def _validate_section_contains(
+    output: str,
+    section_name: str,
+    columns: list[str],
+    message_template: str,
+    violations: list[str],
+) -> None:
+    if f"## {section_name}" not in output:
+        return
+
+    section = re.search(rf"## {section_name}\n(.*?)(?=\n##|\n$)", output, re.DOTALL)
+    if not section:
+        return
+
+    table_content = section.group(1)
+    for column in columns:
+        if column not in table_content:
+            violations.append(message_template.format(column=column))
+
+
+def _validate_csv_parameter_rows(rows: list[list[str]], violations: list[str]) -> None:
+    for row in rows:
+        if len(row) >= 4 and row[3] and not validate_csv_parameters(row[3]):
+            violations.append(f"CSV format invalid parameter encoding: {row[3]}")
+
+
+def _append_extra_compact_entries(
+    violations: list[str],
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    key: str,
+    label: str,
+) -> None:
+    compact_names = {row[0] for row in compact_info.get(key, []) if row}
+    full_names = {row[0] for row in full_info.get(key, []) if row}
+    extra_entries = compact_names - full_names
+    if extra_entries:
+        violations.append(
+            f"Compact format has {label} not in full format: {extra_entries}"
+        )
+
+
+def _is_valid_csv_parameter(param: str) -> bool:
+    parts = param.strip().split(":")
+    if len(parts) != 2:
+        return False
+
+    name, type_name = parts
+    return bool(name.strip() and type_name.strip())
+
+
+@dataclass(frozen=True)
+class _NamedEntrySpec:
+    singular: str
+    plural: str
+    key: str
+
+
+@dataclass(frozen=True)
+class _NamedEntryNames:
+    full: set[str]
+    compact: set[str]
+    csv: set[str]
+
+
+def parse_markdown_table(table_content: str) -> list[list[str]]:
+    """Parse Markdown table content into rows."""
+    lines = [line.strip() for line in table_content.strip().split("\n") if line.strip()]
+    if len(lines) < 2:
+        return []
+
+    return [
+        [cell.strip() for cell in line.split("|")[1:-1]]
+        for line in lines[2:]
+        if line.startswith("|") and line.endswith("|")
+    ]
+
+
+def extract_full_format_info(output: str) -> dict[str, Any]:
+    """Extract structured information from full format."""
+    info = {
+        "class_name": "",
+        "package": "",
+        "methods": [],
+        "fields": [],
+        "imports": [],
+    }
+    _add_full_header_info(output, info)
+    info["methods"] = _extract_markdown_section_rows(output, "Methods")
+    info["fields"] = _extract_markdown_section_rows(output, "Fields")
+    info["imports"] = _extract_import_rows(output)
+    return info
+
+
+def extract_compact_format_info(output: str) -> dict[str, Any]:
+    """Extract structured information from compact format."""
+    info = {"class_name": "", "methods": [], "fields": [], "counts": {}}
+    header_match = re.search(r"^# (.+)$", output, re.MULTILINE)
+    if header_match:
+        info["class_name"] = header_match.group(1)
+
+    _add_compact_counts(output, info)
+    info["methods"] = _extract_markdown_section_rows(output, "Methods")
+    info["fields"] = _extract_markdown_section_rows(output, "Fields")
+    return info
+
+
+def extract_csv_format_info(output: str) -> dict[str, Any]:
+    """Extract structured information from CSV format."""
+    info = {"class_name": "", "methods": [], "fields": [], "classes": []}
+    try:
+        reader = csv.reader(io.StringIO(output))
+        header = next(reader)
+    except (csv.Error, StopIteration):
+        return info
+
+    for row in reader:
+        if len(row) >= 8:
+            _add_csv_row(info, dict(zip(header, row, strict=False)))
+    return info
+
+
+def validate_method_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    violations: list[str],
+) -> bool:
+    """Validate method information consistency."""
+    return _validate_named_entry_consistency(
+        _NamedEntrySpec("Method", "Methods", "methods"),
+        full_info,
+        compact_info,
+        csv_info,
+        violations,
+    )
+
+
+def validate_field_consistency(
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    violations: list[str],
+) -> bool:
+    """Validate field information consistency."""
+    return _validate_named_entry_consistency(
+        _NamedEntrySpec("Field", "Fields", "fields"),
+        full_info,
+        compact_info,
+        csv_info,
+        violations,
+    )
+
+
+def _add_full_header_info(output: str, info: dict[str, Any]) -> None:
+    header_match = re.search(r"^# (.+)$", output, re.MULTILINE)
+    if not header_match:
+        return
+
+    full_name = header_match.group(1)
+    if "." in full_name:
+        parts = full_name.split(".")
+        info["class_name"] = parts[-1]
+        info["package"] = ".".join(parts[:-1])
+        return
+
+    info["class_name"] = full_name
+
+
+def _extract_markdown_section_rows(output: str, section_name: str) -> list[list[str]]:
+    section = re.search(
+        rf"## {re.escape(section_name)}\s*\n(.*?)(?=\n## |\Z)",
+        output,
+        re.DOTALL,
+    )
+    return parse_markdown_table(section.group(1)) if section else []
+
+
+def _extract_import_rows(output: str) -> list[list[str]]:
+    imports_section = re.search(r"## Imports\s*\n(.*?)(?=\n## |\Z)", output, re.DOTALL)
+    if not imports_section:
+        return []
+
+    content = imports_section.group(1)
+    if "```" not in content:
+        return parse_markdown_table(content)
+
+    code_match = re.search(r"```\w*\n(.*?)\n```", content, re.DOTALL)
+    if not code_match:
+        return []
+
+    import_lines = code_match.group(1).strip().split("\n")
+    return [[line.strip()] for line in import_lines if line.strip()]
+
+
+def _add_compact_counts(output: str, info: dict[str, Any]) -> None:
+    info_section = re.search(r"## Info\n(.*?)\n\n", output, re.DOTALL)
+    if not info_section:
+        return
+
+    for row in parse_markdown_table(info_section.group(1)):
+        if len(row) < 2:
+            continue
+        property_name = row[0].lower()
+        if property_name not in ["methods", "fields"]:
+            continue
+        try:
+            info["counts"][property_name] = int(row[1])
+        except ValueError:
+            pass
+
+
+def _add_csv_row(info: dict[str, Any], row_dict: dict[str, str]) -> None:
+    row_type = row_dict.get("Type", "")
+    if row_type == "class":
+        info["classes"].append(row_dict)
+        class_name = row_dict.get("Name", "")
+        info["class_name"] = (
+            class_name.split(".")[-1] if "." in class_name else class_name
+        )
+    elif row_type in ["method", "constructor"]:
+        info["methods"].append(row_dict)
+    elif row_type in ["field", "property"]:
+        info["fields"].append(row_dict)
+
+
+def _validate_named_entry_consistency(
+    spec: _NamedEntrySpec,
+    full_info: dict[str, Any],
+    compact_info: dict[str, Any],
+    csv_info: dict[str, Any],
+    violations: list[str],
+) -> bool:
+    names = _NamedEntryNames(
+        full=_markdown_row_names(full_info.get(spec.key, [])),
+        compact=_markdown_row_names(compact_info.get(spec.key, [])),
+        csv={row.get("Name", "") for row in csv_info.get(spec.key, [])},
+    )
+    names.csv.discard("")
+
+    _append_count_mismatches(spec, names, violations)
+    _append_name_mismatches(spec, names, violations)
+    return True
+
+
+def _markdown_row_names(rows: list[list[str]]) -> set[str]:
+    names = {row[0] for row in rows if row}
+    names.discard("")
+    return names
+
+
+def _append_count_mismatches(
+    spec: _NamedEntrySpec, names: _NamedEntryNames, violations: list[str]
+) -> None:
+    if len(names.full) != len(names.compact):
+        violations.append(
+            f"{spec.singular} count mismatch: "
+            f"full={len(names.full)}, compact={len(names.compact)}"
+        )
+
+    if len(names.full) != len(names.csv):
+        violations.append(
+            f"{spec.singular} count mismatch: "
+            f"full={len(names.full)}, csv={len(names.csv)}"
+        )
+
+
+def _append_name_mismatches(
+    spec: _NamedEntrySpec, names: _NamedEntryNames, violations: list[str]
+) -> None:
+    if names.full and names.compact:
+        _append_set_difference(
+            violations,
+            names.full - names.compact,
+            f"{spec.plural} missing in compact format",
+        )
+        _append_set_difference(
+            violations,
+            names.compact - names.full,
+            f"Extra {spec.key} in compact format",
+        )
+
+    if names.full and names.csv:
+        _append_set_difference(
+            violations,
+            names.full - names.csv,
+            f"{spec.plural} missing in CSV format",
+        )
+        _append_set_difference(
+            violations, names.csv - names.full, f"Extra {spec.key} in CSV format"
+        )
+
+
+def _append_set_difference(
+    violations: list[str], difference: set[str], message: str
+) -> None:
+    if difference:
+        violations.append(f"{message}: {difference}")
+
+
+CONTRACT_CLASS_NAME = "ContractTestService"
+
+CONTRACT_TEST_SERVICE_JAVA_CONTENT = """package com.example.contracts;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+
+/**
+ * Contract testing service
+ * Demonstrates comprehensive class structure for format contract validation
+ */
+public class ContractTestService {
+
+    // Static fields
+    private static final String SERVICE_NAME = "ContractTestService";
+    private static final int MAX_RETRIES = 3;
+
+    // Instance fields
+    private final Map<String, Object> configuration;
+    private List<String> activeConnections;
+    private boolean isInitialized = false;
+
+    /**
+     * Default constructor
+     */
+    public ContractTestService() {
+        this.configuration = new HashMap<>();
+        this.activeConnections = new ArrayList<>();
+    }
+
+    /**
+     * Constructor with configuration
+     * @param config initial configuration
+     */
+    public ContractTestService(Map<String, Object> config) {
+        this.configuration = new HashMap<>(config);
+        this.activeConnections = new ArrayList<>();
+        this.isInitialized = true;
+    }
+
+    /**
+     * Initialize the service
+     * @param timeout timeout in milliseconds
+     * @return initialization result
+     */
+    public CompletableFuture<Boolean> initialize(long timeout) {
+        return CompletableFuture.supplyAsync(() -> {
+            // Initialization logic
+            this.isInitialized = true;
+            return true;
+        });
+    }
+
+    /**
+     * Process data with multiple parameters
+     * @param data input data list
+     * @param options processing options
+     * @param callback completion callback
+     * @return processing result
+     */
+    public Optional<String> processData(
+        List<String> data,
+        Map<String, Object> options,
+        Runnable callback
+    ) {
+        if (!isInitialized) {
+            return Optional.empty();
+        }
+
+        // Processing logic
+        callback.run();
+        return Optional.of("processed");
+    }
+
+    /**
+     * Get service configuration
+     * @return configuration map
+     */
+    public Map<String, Object> getConfiguration() {
+        return new HashMap<>(configuration);
+    }
+
+    /**
+     * Update configuration
+     * @param key configuration key
+     * @param value configuration value
+     */
+    public void updateConfiguration(String key, Object value) {
+        configuration.put(key, value);
+    }
+
+    /**
+     * Check if service is initialized
+     * @return true if initialized
+     */
+    public boolean isInitialized() {
+        return isInitialized;
+    }
+
+    /**
+     * Get active connections count
+     * @return number of active connections
+     */
+    public int getActiveConnectionsCount() {
+        return activeConnections.size();
+    }
+
+    /**
+     * Add connection
+     * @param connectionId connection identifier
+     */
+    public void addConnection(String connectionId) {
+        if (!activeConnections.contains(connectionId)) {
+            activeConnections.add(connectionId);
+        }
+    }
+
+    /**
+     * Remove connection
+     * @param connectionId connection identifier
+     * @return true if connection was removed
+     */
+    public boolean removeConnection(String connectionId) {
+        return activeConnections.remove(connectionId);
+    }
+
+    /**
+     * Shutdown the service
+     */
+    public void shutdown() {
+        activeConnections.clear();
+        configuration.clear();
+        isInitialized = false;
+    }
+}
+"""
+
+
+def create_comprehensive_contract_fixture() -> tuple[str, Path, str]:
+    """Create the Java source file used by format contract tests."""
+    temp_dir = tempfile.mkdtemp()
+    test_file = Path(temp_dir) / f"{CONTRACT_CLASS_NAME}.java"
+    test_file.write_text(CONTRACT_TEST_SERVICE_JAVA_CONTENT, encoding="utf-8")
+    return temp_dir, test_file, CONTRACT_CLASS_NAME
+
+
+def cleanup_comprehensive_contract_fixture(temp_dir: str, test_file: Path) -> None:
+    """Remove the Java source fixture and temporary directory."""
+    test_file.unlink()
+    Path(temp_dir).rmdir()
 
 
 class FormatContractValidator:
@@ -123,10 +982,6 @@ class FormatContractValidator:
     def _extract_csv_format_info(self, output: str) -> dict[str, Any]:
         """Extract structured information from CSV format"""
         return extract_csv_format_info(output)
-
-    def _parse_markdown_table(self, table_content: str) -> list[list[str]]:
-        """Parse Markdown table content into rows"""
-        return parse_markdown_table(table_content)
 
     def _validate_class_consistency(
         self,
@@ -336,7 +1191,9 @@ class TestFormatContracts:
         outputs = await collect_format_outputs(tool, file_path, format_types=("csv",))
         assert_csv_parameter_encoding(outputs["csv"], contract_validator)
         assert "csv" in outputs
-        assert len(outputs["csv"]) > 0
+        assert isinstance(outputs["csv"], str)
+        # CSV output contains at least a header row (non-empty string)
+        assert len(outputs["csv"].strip()) > 0  # ratchet: nondeterministic formatter output length varies by fixture
 
     @pytest.mark.asyncio
     async def test_access_modifier_consistency_contract(
@@ -351,9 +1208,9 @@ class TestFormatContracts:
             contract_validator, outputs
         )
         assert_access_modifier_consistency(full_info, compact_info, csv_info)
-        assert full_info is not None
-        assert compact_info is not None
-        assert csv_info is not None
+        assert isinstance(full_info, dict)
+        assert isinstance(compact_info, dict)
+        assert isinstance(csv_info, dict)
 
     @pytest.mark.asyncio
     async def test_line_number_consistency_contract(
